+++ tests/interface/accessors/index.ts
@@ -1,0 +1,10 @@
+export interface Foo {
+
+    get bar(): string;
+    set bar(value: string);
+
+    get baz(): string;
+
+    set value(value: string);
+
+}