+++ tests/interface/basic/index.ts
@@ -1,0 +1,10 @@
+export interface Foo<K> {
+    x: number;
+    y?: string | K;
+    bar: () => void;
+    isBar: (a: number, b: string) => boolean;
+    checkBar(): Promise<void>;
+    barWithTypedParameter<T>(c: T): T;
+    arrowFuncWithTypedParameters: <T>(c: T) => T;
+    readonly z: number;
+}