+++ tests/interface/index-signature/index.ts
@@ -1,0 +1,9 @@
+export interface StringArray {
+    [index: number]: string;
+}
+
+export interface NumberOrStringDictionary {
+    [index: string]: number | string;
+    length: number;
+    name: string;
+}