+++ packages/core/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Reflect imports with side effects (example: `import './foo.js'`)
 * Imports with different kinds weren't reflected correctly (example: `import Foo, { bar } from './foo'`)
+* The parameter type and name of an index signature wasn't being reflected correctly
 
 ### 🚨 Breaking Changes